# spec/openapi.yaml
openapi: 3.1.0
info:
  title: DeFi Dashboard API
  version: 1.0.0
  description: Backend API for DeFi portfolio tracking and management
  contact:
    name: API Support
    email: api@defi-dashboard.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.defi-dashboard.com/v1
    description: Production server
  - url: http://localhost:3000/v1
    description: Development server
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
    page:
      name: page
      in: query
      description: Page number for pagination
      schema:
        type: integer
        minimum: 1
        default: 1
    limit:
      name: limit
      in: query
      description: Number of items per page
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
...
    address:
      name: address
      in: path
      description: Ethereum address
      required: true
      schema:
        type: string
        pattern: '^0x[a-fA-F0-9]{40}$'
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        details:
          type: object
      required:
        - code
        - message
    PaginationMeta:
      type: object
      properties:
        page:
          type: integer
        limit:
          type: integer
        total:
          type: integer
        totalPages:
          type: integer
...
    PnLExport:
      type: object
      properties:
        startDate:
          type: string
          format: date
        endDate:
          type: string
          format: date
        totalPnl:
          type: number
        totalPnlPercentage:
          type: number
        transactions:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              type:
                type: string
              asset:
                type: string
              amount:
                type: number
              price:
                type: number
              pnl:
                type: number
              fees:
                type: number
    Alert:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [price_above, price_below, large_transfer, approval, liquidity_change]
        status:
          type: string
          enum: [active, triggered, expired, disabled]
        target:
          type: object
          properties:
            type:
              type: string
              enum: [token, address, pool]
            identifier:
              type: string
            chainId:
              type: integer
        conditions:
          type: object
        notification:
          type: object
          properties:
            email:
              type: boolean
            webhook:
              type: string
        createdAt:
          type: string
          format: date-time
        triggeredAt:
          type: string
          format: date-time
      required:
        - id
        - type
        - status
        - target
        - conditions
        - createdAt
...
security:
  - bearerAuth: []
paths:
...
  /analytics/export/{address}:
    get:
      tags:
        - analytics
      summary: Export P&L report
      operationId: exportPnL
      parameters:
        - $ref: '#/components/parameters/address'
        - name: startDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: endDate
          in: query
          required: true
          schema:
            type: string
            format: date
        - name: format
          in: query
          schema:
            type: string
            enum: [json, csv]
            default: json
      responses:
        '200':
          description: P&L report
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PnLExport'
            text/csv:
              schema:
                type: string
  /alerts:
    get:
      tags:
        - alerts
      summary: Get user alerts
      operationId: getAlerts
      parameters:
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/limit'
        - name: status
          in: query
          schema:
            type: string
            enum: [active, triggered, expired, disabled]
      responses:
        '200':
          description: User alerts
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/Alert'
                  meta:
                    $ref: '#/components/schemas/PaginationMeta'

    post:
      tags:
        - alerts
      summary: Create alert
      operationId: createAlert
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                type:
                  type: string
                  enum: [price_above, price_below, large_transfer, approval, liquidity_change]
                target:
                  type: object
                  properties:
                    type:
                      type: string
                      enum: [token, address, pool]
                    identifier:
                      type: string
                    chainId:
                      type: integer
                conditions:
                  type: object
                notification:
                  type: object
                  properties:
                    email:
                      type: boolean
                    webhook:
                      type: string
              required:
                - type
                - target
                - conditions
      responses:
        '201':
          description: Alert created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '400':
          $ref: '#/components/responses/BadRequest'
  /alerts/{alertId}:
    get:
      tags:
        - alerts
      summary: Get alert by ID
      operationId: getAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Alert details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      tags:
        - alerts
      summary: Update alert
      operationId: updateAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, disabled]
                conditions:
                  type: object
                notification:
                  type: object
      responses:
        '200':
          description: Alert updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Alert'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      tags:
        - alerts
      summary: Delete alert
      operationId: deleteAlert
      parameters:
        - name: alertId
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Alert deleted
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  responses:
    BadRequest:
      description: Bad Request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "VALIDATION_ERROR"
            message: "Invalid request parameters"
            details:
              field: "address"
              issue: "Invalid Ethereum address format"
...
    NotFound:
      description: Not Found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            code: "NOT_FOUND"
            message: "Resource not found"
...